# Use asyncio to do requests in parallel