Use async/await to perform all requests in parallel instead of serially.

This adds a dependency on httpx.

This is much faster on the raspberry pi:

Before:
```
hsandsmark@Hakons-MacBook-Pro met.api.no-proxy % curl -w "@curl-format.txt" "https://api.skredkart.no/?lat=60.1234&lon=9.9925"    
{"longitude": 9.9925, "latitude": 60.1234, "radar_coverage": true, "precipitation_amount": 0.0, "precipitation_rate": [[1740952800, 0.0], [1740953100, 0.0], [1740953400, 0.0], [1740953700, 0.0], [1740954000, 0.0], [1740954300, 0.0], [1740954600, 0.0], [1740954900, 0.0], [1740955200, 0.0], [1740955500, 0.0], [1740955800, 0.0], [1740956100, 0.0], [1740956400, 0.0], [1740956700, 0.0], [1740957000, 0.0], [1740957300, 0.0], [1740957600, 0.0], [1740957900, 0.0], [1740958200, 0.0], [1740958500, 0.0], [1740958800, 0.0], [1740959100, 0.0], [1740959400, 0.0]], "air_temperature": 1.1, "relative_humidity": 90.4, "wind_from_direction": 274.9, "wind_speed": 1.2, "symbol_code": "partlycloudy_night", "location_name": "Lat, Lon: 60.1234, 9.9925", "warning_icon": null}     time_namelookup:  0.003453s
        time_connect:  0.006143s
     time_appconnect:  0.030509s
    time_pretransfer:  0.030566s
       time_redirect:  0.000000s
  time_starttransfer:  6.424258s
                     ----------
          time_total:  6.424312s
```

After:
```
hsandsmark@Hakons-MacBook-Pro met.api.no-proxy % curl -w "@curl-format.txt" "https://api.skredkart.no/?lat=60.1234&lon=9.63426"
{"longitude": 9.63426, "latitude": 60.1234, "radar_coverage": true, "precipitation_amount": 0.0, "precipitation_rate": [[1740952800, 0.0], [1740953100, 0.0], [1740953400, 0.0], [1740953700, 0.0], [1740954000, 0.0], [1740954300, 0.0], [1740954600, 0.0], [1740954900, 0.0], [1740955200, 0.0], [1740955500, 0.0], [1740955800, 0.0], [1740956100, 0.0], [1740956400, 0.0], [1740956700, 0.0], [1740957000, 0.0], [1740957300, 0.0], [1740957600, 0.0], [1740957900, 0.0], [1740958200, 0.0], [1740958500, 0.0], [1740958800, 0.0], [1740959100, 0.0], [1740959400, 0.0]], "air_temperature": 4.1, "relative_humidity": 75.0, "wind_from_direction": 249.6, "wind_speed": 2.3, "symbol_code": "partlycloudy_night", "location_name": "Lat, Lon: 60.1234, 9.63426", "warning_icon": null}     time_namelookup:  0.002734s
        time_connect:  0.005544s
     time_appconnect:  0.030375s
    time_pretransfer:  0.030429s
       time_redirect:  0.000000s
  time_starttransfer:  0.935310s
                     ----------
          time_total:  0.935348s